[state_notify] Don't execute any GCode during the 'klipper:ready' event
Klipper commits d1c0cbd63ab2 ("reactor: Add support for temporarily
disabling reactor pauses") and 302df255db41 ("klippy: Verify nothing
attempts to pause in the klippy:ready event") disallowed running any
GCode during the 'klipper:ready' event. (For further information, see
this post: https://klipper.discourse.group/t/prevention-of-reactor-pause/25400).

The 'state_notify' extension does run GCode during the event as part
of the 'on_ready' handling. In order to avoid Klipper erroring, don't
run the gcode as part of the 'klipper:ready' callback.

Instaead, register a callback that is called, which registers a reactor
callback with the wake up time of 'NOW'. The result of this is that
the 'on_ready' template is execute virtually immediately, preserving
the intent and order.